requirements.txt
absl-py==0.7.0
aioprocessing==1.0.1
alembic==1.0.7
appnope==0.1.0
argh==0.26.2
astor==0.7.1
astroid==2.2.4
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.202
botocore==1.12.202
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
consoleprinter==93
cycler==0.10.0
Cython==0.29.6
dataset==1.1.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docker-pycreds==0.4.0
docutils==0.14
entrypoints==0.3
filelock==3.0.12
flake8==3.6.0
flatbuffers==1.11
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.1
gql==0.1.0
grpcio==1.19.0
gym==0.15.4
h5py==2.9.0
httplib2==0.13.1
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
importlib-metadata==1.3.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.12
janus==0.4.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lockfile==0.12.2
lz4==2.2.1
matplotlib
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==8.0.2
moviepy==1.0.0
mpi4py==3.0.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.18.0
numpy-stl==2.10.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.4.0
pathtools==0.1.2
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
plotly==3.8.1
pluggy==0.13.1
proglog==0.1.9
progressbar2==3.39.2
prometheus-client==0.6.0
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.11.2
psutil==5.5.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
pyglet==1.3.2
Pygments==2.3.1
pylint==2.3.1
Pympler==0.7
PyOpenGL==3.1.0
pyparsing==2.4.5
pyrsistent==0.15.1
pytest==5.3.2
python-dateutil==2.8.0
python-editor==1.0.4
python-utils==2.3.0
pytz==2019.1
PyYAML==5.2
pyzmq==18.0.0
qtconsole==4.4.3
ray==0.8.0
redis==3.3.11
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
scipy==1.4.1
Send2Trash==1.5.0
sentry-sdk==0.7.4
setproctitle==1.1.10
shortuuid==0.5.0
singledispatch==3.4.0.3
six==1.13.0
slackclient==1.3.1
smmap2==2.0.5
SQLAlchemy==1.2.18
subprocess32==3.5.3
tabulate==0.8.6
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
torch==1.1.0
torchvision==0.2.1
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.3.1
typing==3.7.4.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.1
wandb==0.8.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.0.2
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
yappi==1.0
zipp==0.6.0
